+++ src/command/update_release.rs
@@ -1,13 +1,14 @@
 use crate::common::project::ProjectRef;
 use crate::GitlabAuth;
 use chrono::{DateTime, Utc};
+use compact_str::CompactString;
 
 /// Update a project release
 ///
 /// <https://docs.gitlab.com/ee/api/releases/#update-a-release>
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct UpdateReleaseCommand<Cx, Str = String> {
+pub struct UpdateReleaseCommand<Cx, Str = CompactString> {
   pub context: Cx,
   pub auth: Option<GitlabAuth<Str>>,
   pub project: ProjectRef<Str>,