+++ src/query/get_project_release.rs
@@ -1,12 +1,13 @@
 use crate::common::project::ProjectRef;
 use crate::GitlabAuth;
+use compact_str::CompactString;
 
 /// Get a project release
 ///
 /// <https://docs.gitlab.com/ee/api/releases/#get-a-release-by-a-tag-name>
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct GetProjectReleaseQuery<Cx, Str = String> {
+pub struct GetProjectReleaseQuery<Cx, Str = CompactString> {
   pub context: Cx,
   pub auth: Option<GitlabAuth<Str>>,
   pub project: ProjectRef<Str>,